+++ packages/react/src/components/TextInput/styles.ts
@@ -5,9 +5,19 @@ export const TextInputContainer = styled('div', {
   borderRadius: '$sm',
   boxSizing: 'border-box',
   border: '2px solid $gray900',
-
   display: 'flex',
   alignItems: 'center',
+
+  variants: {
+    size: {
+      sm: {
+        padding: '$2 $3',
+      },
+      md: {
+        padding: '$3 $4',
+      },
+    },
+  },
 
   '&:has(input:focus)': {
     borderColor: '$ignite300',
@@ -16,13 +26,6 @@ export const TextInputContainer = styled('div', {
   '&:has(input:disabled)': {
     opacity: 0.5,
     cursor: 'not-allowed',
-  },
-
-  variants: {
-    size: {
-      sm: { padding: '$2 $3' },
-      md: { padding: '$3 $4' },
-    },
   },
 
   defaultVariants: {
@@ -34,15 +37,15 @@ export const Prefix = styled('span', {
   fontFamily: '$default',
   fontSize: '$sm',
   color: '$gray400',
-  fontWeight: '$regular',
+  fontWeight: 'regular',
 })
 
 export const Input = styled('input', {
   fontFamily: '$default',
   fontSize: '$sm',
   color: '$white',
-  fontWeight: '$regular',
-  backgroundColor: 'transparent',
+  fontWeight: 'regular',
+  background: 'transparent',
   border: 0,
   width: '100%',