+++ packages/tokens/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ignight-ui/tokens
 
+## 3.2.0
+
+### Minor Changes
+
+- Adding TextInput size and removing test color.
+
 ## 3.1.0
 
 ### Minor Changes